requirements.txt
pydantic==1.10.0
fastapi==0.99.1
uvicorn==0.22.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.6
redis==4.5.5
celery==5.3.1
flower==2.0.0
kafka-python==2.0.2
pytest==7.4.0
httpx==0.23.3
